+++ ft-plugin/CHANGELOG.md
@@ -1,3 +1,9 @@
+# plugin 1.3.6-alpha03
+1. Removed ASM for Activity and Fragment.
+2. Added code line number display for Exceptions.
+3. Using `org.gradle.api.logging.Logger` to print log instead of `system.out`
+
+---
 # plugin 1.3.6-alpha02
 1. Added verboseLog param for more logs.
 2. Added try-catch and supplemented error logs.